+++ bootique-rabbitmq-client/src/main/java/io/bootique/rabbitmq/client/pubsub/RmqSubEndpointFactory.java
@@ -21,6 +21,7 @@ package io.bootique.rabbitmq.client.pubsub;
 import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
 import io.bootique.rabbitmq.client.ChannelFactory;
+import io.bootique.shutdown.ShutdownManager;
 
 import java.util.Objects;
 
@@ -36,9 +37,11 @@ public class RmqSubEndpointFactory {
     private String routingKey;
     private boolean autoAck = true;
 
-    public RmqSubEndpoint create(ChannelFactory channelFactory) {
+    public RmqSubEndpoint create(ChannelFactory channelFactory, ShutdownManager shutdownManager) {
         Objects.requireNonNull(connection, "Subscriber connection name is undefined");
-        return new RmqSubEndpoint(channelFactory, connection, exchange, queue, routingKey, autoAck);
+        RmqSubEndpoint endpoint = new RmqSubEndpoint(channelFactory, connection, exchange, queue, routingKey, autoAck);
+        shutdownManager.addShutdownHook(() -> endpoint.close());
+        return endpoint;
     }
 
     @BQConfigProperty