+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 2.0.B1
 
+* #4 Configurable Publishers / Consumers
 * #7 Upgrade Testcontainers to 1.15.0-rc2
 * #8 ChannelFactory: clear unambiguous API
 * #10 Upgrade RMQ client to 5.12.0